Aplicatie_sesizari/requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
pytest==8.3.5
waitress==3.0.0
pywebpush==2.0.3
